+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.10.1] - Unreleased
 ### Fixed
 - Fix interface validation: object aggregated properties interfaces are not valid.
+- Fix interface validation: server owned object aggregated interfaces are not yet supported, hence not valid.
 
 ## [0.10.0] - 2019-04-16